requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
autocommand==2.2.2
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
-e git+https://github.com/apple/ml-depth-pro.git@b2cd0d51daa95e49277a9f642f7fd736b7f9e91d#egg=depth_pro
easydict==1.13
faiss-gpu==1.7.2
filelock @ file:///croot/filelock_1700591183607/work
filterpy==1.4.5
flake8==7.1.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
future==1.0.0
gdown==5.2.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
grpcio==1.67.0
h5py==3.12.1
huggingface-hub==0.26.1
humanfriendly==10.0
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
imageio==2.36.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iopath==0.1.10
isort==4.3.21
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
kiwisolver==1.4.7
loguru==0.7.2
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
mccabe==0.7.0
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
more-itertools==10.3.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1717597493534/work
ninja==1.11.1.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.12.0
onnx-simplifier==0.3.5
onnxoptimizer==0.3.13
onnxruntime==1.12.0
opencv-python==4.10.0.84
packaging==24.1
pillow @ file:///croot/pillow_1721059439630/work
pillow_heif==0.20.0
platformdirs==4.2.2
portalocker==2.10.1
prettytable==3.11.0
protobuf==3.20.1
pycocotools==2.0.8
pycodestyle==2.12.1
pyflakes==3.2.0
pyparsing==3.2.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests @ file:///croot/requests_1721410876868/work
safetensors==0.4.5
-e git+https://github.com/facebookresearch/segment-anything-2.git@7e1596c0b6462eb1d1ba7e1492430fed95023598#egg=SAM_2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smplx==0.1.28
soupsieve==2.6
sympy @ file:///croot/sympy_1724938189289/work
tabulate==0.9.0
tb-nightly==2.19.0a20241021
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==1.0.11
tomli==2.0.1
torch==2.3.1
torchaudio==2.3.1
-e git+https://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@github.com/Milad-Khanchi/VisionSeg-OC-SORT@188c965a502551f2aa7f9366e89aaad1d64519a4#egg=torchreid&subdirectory=external/deep-person-reid
torchvision==0.18.1
tqdm==4.66.5
triton==2.3.1
typeguard==4.3.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth==0.2.13
Werkzeug==3.0.4
yacs==0.1.8
yapf==0.40.2
-e git+https://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@github.com/Milad-Khanchi/VisionSeg-OC-SORT@188c965a502551f2aa7f9366e89aaad1d64519a4#egg=yolox&subdirectory=external/YOLOX
zipp==3.19.2
